feat: mood keyword 타입선언 파일 분리
